+++ example/sirius_example.dart
@@ -1,27 +1,28 @@
-// import 'package:sirius_backend/sirius_backend.dart';
+import 'package:sirius_backend/sirius_backend.dart';
 
-// void main() {
-//   Sirius sirius = Sirius();
+void main() {
+  Sirius sirius = Sirius();
 
-//   sirius.post("api/partner/register", (Request request) async {
-//     Map<String, ValidationRules> rules = {
-//       "kitchen_type": ValidationRules(
-//           required: required(),
-//           inList: inList(["HOME_KITCHEN", "COMMERCIAL_KITCHEN"]))
-//     };
+  sirius.post("test", (Request request) async {
+    Map<String, ValidationRules> rules = {
+      "address": ValidationRules(
+          dataType: dataType(DataTypes.NUMBER),
+          nullable: false,
+          required: required())
+    };
 
-//     Validator validator = Validator(request, rules);
+    Validator validator = Validator(request.getAllFields(), rules);
 
-//     if (!validator.validate()) {
-//       return Response.send(validator.getError.value);
-//     }
+    if (!validator.validate()) {
+      return Response.send(validator.getAllErrors);
+    }
 
-//     return Response.send("Success..!");
-//   });
+    return Response.send("Success..!");
+  });
 
-//   sirius.start(
-//     callback: (server) {
-//       print("server is running");
-//     },
-//   );
-// }
+  sirius.start(
+    callback: (server) {
+      print("server is running");
+    },
+  );
+}